Added MTCNN in the pipeline
